tools/thumbnailer-rust/requirements.txt
Pillow>=9.0.0
scikit-image>=0.20.0
numpy>=1.23.0